requirements.txt
google-generativeai>=0.8.3
streamlit>=1.39.0
pydantic>=2.9.0
pypdf>=5.1.0
python-dotenv>=1.0.1
pandas>=2.0.0
